chore(models): keep package file-count baseline
Consolidate the resolver contracts and focused resolver tests into existing Models contract and service files so the exact package-file-count baseline remains satisfied.
